requirements.txt
finnhub-python
python-dotenv
fastapi
uvicorn[standard]